+++ Baidu Locator API.py
@@ -89,8 +89,9 @@ def Process(apikey, city, city_code):
     f_in.close()              
     f_out.close()
     f_err.close()
-    os.remove("temp.txt")
     
+    if os.path.exists("temp.txt"):
+        os.remove("temp.txt")    
     if err == 0:
         os.remove("Output/Locating Error.csv")
